Move integration files under custom_components
