+++ custom_components/tado_x/__init__.py
@@ -1,1 +1,20 @@
-"""Tado X integration."""
+"""Tado X integration setup."""
+from __future__ import annotations
+
+from homeassistant.config_entries import ConfigEntry
+from homeassistant.core import HomeAssistant
+
+from .const import DOMAIN
+
+
+async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
+    """Set up Tado X from a config entry."""
+    hass.data.setdefault(DOMAIN, {})
+    hass.data[DOMAIN][entry.entry_id] = entry.data
+    return True
+
+
+async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
+    """Unload a Tado X config entry."""
+    hass.data[DOMAIN].pop(entry.entry_id, None)
+    return True
