+++ custom_components/tado_x/manifest.json
